do not include output as per /about#faq_copying
https://adventofcode.com/about#faq_copying
Can I copy/redistribute part of Advent of Code? Please don't. Advent of
Code is free to use, not free to copy. If you're posting a code repository
somewhere, please don't include parts of Advent of Code like the puzzle
text or your inputs. If you're making a website, please don't make it look
like Advent of Code or name it something similar.